RDKB-9150: parental control Managed Devices
Reason for change: WebGUI - GET/SET call for parental control Managed Devices
If user tries to edit a rule to make it a conflicting rule with another existing rule, then a relevant error popup should be shown to user and changes shouldn't take effect.
Test Procedure: Please test and see if GUI is working as expected for,
For EDIT case, If user tries to edit a rule to make it a conflicting rule with another existing rule, then a relevant error popup should be shown to user and changes shouldn't take effect. Only non-conflicting rules should get saved.
Risks: None

Change-Id: I2cf9ea03fd52edcec478f087a1eac1b50e473bf1
Signed-off-by: Pavan Kumar PavanKumarReddy_B@infosys.com