RDKB-8944: Validation against XSS for Advanced sub menu
Reason for change: Validation of the input parameter(s) to prevent WebUI - Cross-Site Scripting in Advanced sub menu
Validation of the input parameter(s) is implemented on back-end of GUI page in "actionHandler" folder.
Test Procedure: Please test and see if GUI is working as expected for,
Use the "Burp Suite" to intercept the requests mentioned below to modify and check, modified invalid requests shouldn't take effect on the XB-3 device.
Risks: None

Change-Id: Iea745a8eafaf625bcd4ff1666c2d62562e4a5fd3
Signed-off-by: Pavan Kumar PavanKumarReddy_B@infosys.com