RDKB-9506: prevent XSS for Captive Portal
Reason for change: Validation of the input parameter(s) to prevent WebUI - Cross-Site Scripting for Captive Portal
Validation of the input parameter(s) is implemented on back-end of GUI page in "actionHandler" folder.
Test Procedure: Please test and see if GUI is working as expected for Captive Portal
Use the "Burp Suite" to intercept the requests mentioned below to modify and check, modified invalid requests shouldn't take effect on the XB-3 device.
Risks: None

Change-Id: I3aca04ddde9d33e408169f4554cadc822040376c
Signed-off-by: Pavan Kumar PavanKumarReddy_B@infosys.com