RDKB-9480: prevent XSS in At a Glance & Email Notification pages
Reason for change: Validation of the input parameter(s) to prevent WebUI - Cross-Site Scripting in At a Glance & Email Notification pages
Validation of the input parameter(s) is implemented on back-end of GUI page in "actionHandler" folder.
Test Procedure: Please test and see if GUI is working as expected for At a Glance & Email Notification pages
Use the "Burp Suite" to intercept the requests mentioned below to modify and check, modified invalid requests shouldn't take effect on the XB-3 device.
Risks: None

Change-Id: If4fb9b222ad309bf9304ac1b02af080df7665adf
Signed-off-by: Pavan Kumar PavanKumarReddy_B@infosys.com