SKYH4-2650,SKYH4-2669: ULA prefix value set to 0.0.0.0 after saving IPV6 .
Reason for change: added new conditional check for empty and vaue set to 0 is invalid.

Test Procedure: Login to GUI and navigate to 'Connessione >Rete Local' Try setting the Prefix ULA address to 0 or empty.

Risks: low

Signed-off-by: Viswa Patneedi <viswa.patneedi@sky.uk>

SKYH4-2669: ULA prefix as 0000 on any of the fields except the 1st field.

Reason for change: Removed the additional conditional check for 0 value .

Test Procedure: Login to GUI and navigate to 'Connessione >Rete Local' Try setting the Prefix ULA address to 0 or empty.

Risks: low

Change-Id: I97ed69cd0ac34fa8dc7f738c56b2d35dc67008d5
Signed-off-by: Viswa Patneedi <viswa.patneedi@sky.uk>